requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
gspread==5.12.3
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
openai==1.10.0
pydantic==2.6.1
pydantic-settings==2.1.0
schedule==1.2.0
python-multipart==0.0.9
